# state/runs/2026-10-09-040829-umm-improve/validated-snapshot/docs/openapi.yaml
openapi: 3.1.0
info:
  title: umm API
  version: 0.8.1
  description: |
    Spatial Thought Memory, collaboration, automation and Dream Layer API.
    Supported Canvas mutation requests can include an `Idempotency-Key` header (8-128 safe characters) for replay-safe offline synchronization.
    Personal preferences, API-key lifecycle operations and all administrator operations require an interactive browser session; bearer API keys cannot call them.
    Errors use RFC 9457 Problem Details and note version conflicts include the latest server note.
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      operationId: post_auth_login
      security: []
      tags: [Auth]
      summary: Sign in with the local password; concurrent guesses share the configured PostgreSQL-backed ceiling
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [username, password], properties: { username: { type: string }, password: { type: string, format: password } } } } }
      responses:
        '200': { description: Signed in, headers: { Set-Cookie: { schema: { type: string } } }, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '429': { description: login-locked Problem Details with Retry-After after the shared failure ceiling, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '503': { description: Fail-closed response when the durable throttle cannot be checked or committed, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /auth/oidc/start:
    get:
      operationId: get_auth_oidc_start
      security: []
      tags: [Auth]
      summary: Begin Keycloak Authorization Code flow
      description: >-
        return_to must be a site path — starting with one slash, not two — or it becomes /.
        prompt=none asks Keycloak to answer from an existing session without drawing a screen (silent SSO);
        it is honoured only while the administrator has turned auto_login on, and is otherwise treated as an
        ordinary login so that appending it to a URL cannot change the flow.
      parameters:
        - { in: query, name: return_to, schema: { type: string, default: / } }
        - { in: query, name: prompt, schema: { type: string, enum: [none] } }
      responses: { '302': { description: Redirect to Keycloak } }
  /auth/oidc/callback:
    get:
      operationId: get_auth_oidc_callback
      security: []
      tags: [Auth]
      summary: Complete Keycloak Authorization Code flow
      description: >-
        When Keycloak answers with error instead of code the browser is sent to the login screen:
        /login?sso=none for login_required, interaction_required and consent_required (the ordinary reply to a
        silent prompt=none attempt when there is no session — the marker tells the browser not to try again),
        /login?sso=error for anything else.
      parameters:
        - { in: query, name: state, required: true, schema: { type: string } }
        - { in: query, name: code, schema: { type: string } }
        - { in: query, name: error, schema: { type: string } }
      responses: { '302': { description: Session established and redirected to return_to, or refused and redirected to /login } }
...
  /spaces:
    get:
      operationId: get_spaces
      tags: [Thoughts]
      summary: List visible spaces
      x-required-scope: spaces:read
      responses: { '200': { description: Spaces, content: { application/json: { schema: { type: object, properties: { spaces: { type: array, items: { $ref: '#/components/schemas/Space' } } } } } } } }
    post:
      operationId: post_spaces
      tags: [Thoughts]
      summary: Create a space
      x-required-scope: notes:write
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [name], properties: { name: { type: string, maxLength: 200 } } } } } }
      responses: { '201': { description: Space created, content: { application/json: { schema: { $ref: '#/components/schemas/Space' } } } } }
  /search:
    get:
      operationId: get_search
      tags: [Thoughts]
      summary: Hybrid semantic and lexical search across accessible spaces
      x-required-scope: notes:read
      parameters:
        - { in: query, name: q, required: true, schema: { type: string, maxLength: 200 } }
        - { in: query, name: spaceId, schema: { type: string, format: uuid } }
        - { in: query, name: kind, schema: { type: string } }
        - { in: query, name: updatedFrom, schema: { type: string, format: date-time } }
        - { in: query, name: updatedTo, schema: { type: string, format: date-time } }
        - { $ref: '#/components/parameters/Cursor' }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
      responses:
        '200':
          description: Ranked note shortcuts
          content:
            application/json:
              schema:
                type: object
                required: [notes, nextCursor]
                properties:
                  notes: { type: array, items: { $ref: '#/components/schemas/NoteSearchResult' } }
                  nextCursor: { type: string }
                  noteLines:
                    type: object
                    description: >
                      Maps a result to the line of thinking it belongs to, when it belongs to
                      one. A result whose line is 'abandoned' was considered and set aside,
                      with the reason recorded, and must not be presented as current. Search
                      is where a thought is met most often, and an unlabelled result reads
                      exactly like a live one.
                    additionalProperties: { $ref: '#/components/schemas/Branch' }
  /spaces/{spaceId}/notes:
    parameters:
      - { $ref: '#/components/parameters/SpaceID' }
    get:
      operationId: get_spaces_spaceId_notes
      tags: [Thoughts]
      summary: List or search notes and edges in a space
      x-required-scope: notes:read
      parameters:
        - { in: query, name: q, schema: { type: string } }
      responses:
        '200':
          description: Canvas graph
          content: { application/json: { schema: { type: object, properties: { notes: { type: array, items: { $ref: '#/components/schemas/Note' } }, edges: { type: array, items: { $ref: '#/components/schemas/Edge' } } } } } }
    post:
      operationId: post_spaces_spaceId_notes
      tags: [Thoughts]
      summary: Drop a post-it onto the canvas
      x-required-scope: notes:write
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/NoteWrite' } } } }
      responses: { '201': { description: Note created, content: { application/json: { schema: { $ref: '#/components/schemas/Note' } } } } }
  /spaces/{spaceId}:
    parameters:
      - { $ref: '#/components/parameters/SpaceID' }
    put:
      operationId: put_spaces_spaceId
      tags: [Thoughts]
      summary: Rename a space or change its AI exclusion policy
      x-required-scope: notes:write
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [name], properties: { name: { type: string, minLength: 1, maxLength: 200 }, aiExcluded: { type: boolean, description: Exclude every note in this space from AI processing and remote embedding gateways } } } } }
      responses:
        '200': { description: Space updated, content: { application/json: { schema: { $ref: '#/components/schemas/Space' } } } }
    delete:
      operationId: delete_spaces_spaceId
      tags: [Thoughts]
      summary: Permanently delete an owned space and its contents
      description: Only the owner can delete a space. Notes, edges, Dream records and membership data in it are removed by cascade.
      x-required-scope: notes:write
      responses:
        '204': { description: Space deleted }
        '403': { $ref: '#/components/responses/Error' }
  /notes/{noteId}:
    parameters:
      - { $ref: '#/components/parameters/NoteID' }
    put:
      operationId: put_notes_noteId
      tags: [Thoughts]
      summary: Autosave note content, appearance and position
      description: Uses optimistic concurrency. Send the last server version; a stale version returns 409. A caller who can still view the note but no longer edit its space receives a terminal note-read-only 403 Problem Details response.
      x-required-scope: notes:write
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/NoteUpdate' } } } }
      responses:
        '200': { description: Updated note, content: { application/json: { schema: { $ref: '#/components/schemas/Note' } } } }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
    delete:
      operationId: delete_notes_noteId
      tags: [Thoughts]
      summary: Soft-delete a note
      x-required-scope: notes:write
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses: { '204': { description: Deleted } }
  /notes/{noteId}/related:
    parameters:
      - { $ref: '#/components/parameters/NoteID' }
    get:
      operationId: get_notes_noteId_related
      tags: [Thoughts]
      summary: Discover semantically related thoughts with the offline embedding
      x-required-scope: notes:read
      parameters:
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 20, default: 8 } }
      responses:
        '200': { description: Ranked related thoughts, content: { application/json: { schema: { type: object, properties: { related: { type: array, items: { $ref: '#/components/schemas/RelatedNote' } } } } } } }
  /notes/{noteId}/backlinks:
    parameters:
      - { $ref: '#/components/parameters/NoteID' }
    get:
      operationId: get_notes_noteId_backlinks
      tags: [Thoughts]
      summary: List incoming and outgoing thought links
      x-required-scope: notes:read
      responses:
        '200': { description: Backlinks, content: { application/json: { schema: { type: object, required: [backlinks], properties: { backlinks: { type: array, items: { $ref: '#/components/schemas/Backlink' } } } } } } }
  /notes/{noteId}/review:
    parameters:
      - { $ref: '#/components/parameters/NoteID' }
    post:
      operationId: post_notes_noteId_review
      tags: [Review]
      summary: Complete, snooze or pin a thought review
      x-required-scope: notes:write
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, properties: { snoozeDays: { type: integer, minimum: 1, maximum: 365 }, pinned: { type: boolean }, complete: { type: boolean, default: true, description: Set false to change only the pinned state without completing the review. } } } } }
      responses:
        '200': { description: Updated review item, content: { application/json: { schema: { $ref: '#/components/schemas/ReviewItem' } } } }
  /notes/{noteId}/comments:
    parameters:
      - { $ref: '#/components/parameters/NoteID' }
    get:
      operationId: get_notes_noteId_comments
      tags: [Collaboration]
      summary: List note comments and mention metadata
      x-required-scope: notes:read
      responses:
        '200': { description: Comments, content: { application/json: { schema: { type: object, required: [comments], properties: { comments: { type: array, items: { $ref: '#/components/schemas/Comment' } } } } } } }
    post:
      operationId: post_notes_noteId_comments
      tags: [Collaboration]
      summary: Add a comment and notify mentioned space members
      x-required-scope: notes:write
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [body], properties: { body: { type: string, minLength: 1, maxLength: 4000 }, parentId: { type: [string, 'null'], format: uuid } } } } }
      responses:
        '201': { description: Comment created, content: { application/json: { schema: { $ref: '#/components/schemas/Comment' } } } }
  /comments/{commentId}/resolve:
    parameters:
      - { $ref: '#/components/parameters/CommentID' }
    put:
      operationId: put_comments_commentId_resolve
      tags: [Collaboration]
      summary: Resolve or reopen a comment thread; permanently forbidden replays return comment-mutation-forbidden
      x-required-scope: notes:write
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [resolved], properties: { resolved: { type: boolean } } } } }
      responses:
        '200': { description: Updated comment, content: { application/json: { schema: { $ref: '#/components/schemas/Comment' } } } }
        '403': { description: Terminal comment-mutation-forbidden Problem Details when the current session can no longer mutate the comment, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /admin/oidc/test:
    post:
      operationId: post_admin_oidc_test
      tags: [Admin]
      summary: Test OIDC Discovery and client configuration
      responses: { '200': { description: Discovery succeeded }, '400': { $ref: '#/components/responses/Error' } }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: umm_session }
    bearerAuth: { type: http, scheme: bearer, bearerFormat: umm_key }
  parameters:
    SpaceID: { in: path, name: spaceId, required: true, schema: { type: string, format: uuid } }
    NoteID: { in: path, name: noteId, required: true, schema: { type: string, format: uuid } }
...
    Cursor: { in: query, name: cursor, description: Opaque cursor returned as nextCursor, schema: { type: string } }
...
  responses:
...
    Error: { description: RFC 9457 Problem Details, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
...
    User:
      type: object
      required: [id, username, displayName, role, active]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        email: { type: string }
        role: { type: string, enum: [user, team_lead, admin] }
        teamId: { type: [string, 'null'], format: uuid }
        active: { type: boolean }
...